// src/store/categories/mutation-types.ts
export enum CategoriesMutationTypes {
    SET_CATEGORIES = 'SET_CATEGORIES',
    SET_CATEGORY_ID = 'SET_CATEGORY_ID',
    SET_CATEGORY='SET_CATEGORY',
    REMOVE_CATEGORY='REMOVE_CATEGORY',
    CHANGE_CATEGORIES_PAGE = 'CHANGE_CATEGORIES_PAGE',
    CHANGE_CATEGORIES_LOADING_LIST = 'CHANGE_CATEGORIES_LOADING_LIST',
    CHANGE_CATEGORIES_LOADING_ITEM = 'CHANGE_CATEGORIES_LOADING_ITEM',
    SET_CATEGORIES_COUNT='SET_CATEGORIES_COUNT'
}